Change bfs strategy
Instead of using the `parents` dict as a way to mark a title
as visited (and only doing so after finding all the linked titles in its
html), use a separate `seen` hashset that includes all titles that we
have found that have been linked to by some other page. We now
collect all titles that have been seen, not just ones that we've
searched the html of already.

Also rename the `to_visit` deque to `queue` to remove its affect on how
we think about what pages we've "visited" so far.