Remove more complex path-building method
Had been working on a different implementation of `_get_path` that
relied on the `parents` dict using hashsets as the values. The idea was
that a page coule have multiple parents that link to it, and we wouldn't
know which one would actually lead to the shortest path. Different
`_get_path` implementations recursively found the shortest resulting
path from using each of the potential parents, and chose the shortest
one.

This seems to have not made a difference in the tests. Removing the
commented out implementation.