Simplify path building
`_get_path` does not need to return a possible None value.

Can also use a list as the path, append to the end, and reverse, instead
of using a deque. A deque would allow for constant time appending to
the beginning, without needing to reverse at the end. But it needs to be
converted to a list at some point, which requires copying. Instead, the
list can be reversed in place.