Implement backend server and crawler in rust
This is all the code I had saved from working on this in July 2024. I
never pushed the commits from the branch where I worked on this Rust
implementation.

When coming back to this project now, pushing the commits from the
branch failed. Trying to access git files resulted in bus failures. And
at some point in trying to retrieve the code, the files were deleted. I
eventually had to copy the code from the diffs of the deleted files and
paste it all into a newly cloned repo.

So here is that implementation. Fun time.