Collect found urls into a list instead of set
Each url at the front of the queue must be checked if it's been seen
already. This requires checking the hashset/hashmap of items already
seen to avoid searching a page that's already been searched.

Because this is already done before searching a page, collecting new
urls into a set would then only have the benefit of not placing
duplicate urls into the queue. However, this hashing operation is
costly, and it is not common for a page to link to another page multiple
times.