Simplify state; share instances in tests
Only instance variable is now an http session/client to reuse on each
sent request.

The starting title is now passed when beginning to find a path. This
makes sense as it's only relevant for a specific search. This also now
allows reuse of instances, and thus sessions, between searches.

Modify tests to create single instances and reuse in each test:
  - python - pytest fixture
  - rust - a lazylock containing the instance that's initialized once