Update async implementation
Apply changes made to searching behavior since first instance of async
implementation.

--

Handle async task execution of `worker` and `get_linked_titles` methods:

There were a few options. We could have `get_linked_titles` take a
`Client` in its signature, and add that to the `worker` signature to
pass to it. Then, we would clone the `Client` handle and move that into
each spawned task. This uses an Arc internally to track references.

To remove the need to expliciy pass a client around, we could interact
with the Self type always wrapped in an Arc. This way, we'd clone the
Self Arc to pass into `worker`, and each task can access the same client
by a reference to self through their clone of the Arc.

However, since `Client` is clone and the only data in the Self struct is
the client, we can derive Clone on the struct and clone self to pass to
tasks. This will clone the handle to client without needing to use Arc,
and allow access to the client in each task through the cloned self.