Fix failing requests
Were getting a high number of instances of failing to collect linked
titles. Turns out that we were getting http responses saying we had sent
an invadid continuation request. These responses could then not be
decoded into `WikiResponses`.

Part of the continuation fields of paginated results included the
characters '||'. When the `WikiResponse` field for this section was a
serde::Value, decoding failed. Changing this type to a String resolved
this issue.

In addition, using two separate calls to `text` and
`serde_json::from_str` instead of one call to the `json` method allowed
for debugging of the response body and seeing the error details provided
when serde failed to decode it into json. I've kept these calls for
potential easier debugging in the future.