Use barrier on channel implementation
Barrier implementation using a coordinator task.

Very inefficient as the coordinator manually drains each title in the
workers' sending channel and sends to the workers' reciving channel.

Going to attempt to reimplement the async version with vectors, and then
come back to this channel implementation.

- `test_two_hops_1` sometimes does not finish when running full test
suite.
- Finishes in ~1 second when running individually.

- `test_two_hops_2` finishes in about ~7 seconds now (was ~2 seconds
with 5 workers before adding the barrier in this commit).